Add a bunch of unit tests for the dashboard
This exercises many different patterns that we might want to test,
so that there are relevant examples to get inspired by for future
unit tests. The only pattern that is a bit difficult to test
currently is modals, since they are rendered in the OverlayContainer
which is not part of a page's DOM tree (but rather defined in
_app.) Modal dialogs should still be testable in isolation though.

Additionally, this uncovered one bug (the add-on banner would also
be visible with the add-on installed) and a couple of accessibility
issues (primarily missing labels for form fields).